refactor(api): zdieľané jadro jednorazových preview-tokenov (issue 505)
Tri bit-identické moduly (nedostupne/order-merge/customer-contact) delegujú
na nový generický store single-use-preview-tokens.ts (factory, per-wrapper
Map). Verejné signatúry aj správanie ostávajú nezmenené — čistý refaktor.
